Isolate AXI 4 Lite monitor into a common file.
Internal-tag: [#51803]
